refactor: shift menu bars to separate file for discoverabiilty
