+++ apps/studio/src/components/PageEditor/MenuBar/AccordionMenuBar.tsx
@@ -1,0 +1,90 @@
+import { useMemo } from "react"
+import { Editor } from "@tiptap/react"
+import {
+  BiBold,
+  BiItalic,
+  BiListOl,
+  BiListUl,
+  BiStrikethrough,
+  BiUnderline,
+} from "react-icons/bi"
+import { MdSubscript, MdSuperscript } from "react-icons/md"
+
+import { MenuBar, MenuBarEntry } from "./MenuBar"
+
+export const AccordionMenuBar = ({ editor }: { editor: Editor }) => {
+  const items: MenuBarEntry[] = useMemo(
+    () => [
+      {
+        type: "item",
+        icon: BiBold,
+        title: "Bold",
+        action: () => editor.chain().focus().toggleBold().run(),
+        isActive: () => editor.isActive("bold"),
+      },
+      {
+        type: "item",
+        icon: BiItalic,
+        title: "Italicise",
+        action: () => editor.chain().focus().toggleItalic().run(),
+        isActive: () => editor.isActive("italic"),
+      },
+      {
+        type: "item",
+        icon: BiUnderline,
+        title: "Underline",
+        action: () => editor.chain().focus().toggleUnderline().run(),
+        isActive: () => editor.isActive("underline"),
+      },
+      {
+        type: "item",
+        icon: BiStrikethrough,
+        title: "Strikethrough",
+        action: () => editor.chain().focus().toggleStrike().run(),
+        isActive: () => editor.isActive("strike"),
+      },
+      {
+        type: "item",
+        icon: MdSuperscript,
+        title: "Superscript",
+        action: () => editor.chain().focus().toggleSuperscript().run(),
+        isActive: () => editor.isActive("superscript"),
+      },
+      {
+        type: "item",
+        icon: MdSubscript,
+        title: "Subscript",
+        action: () => editor.chain().focus().toggleSubscript().run(),
+        isActive: () => editor.isActive("subscript"),
+      },
+      {
+        type: "divider",
+      },
+      {
+        type: "horizontal-list",
+        label: "Lists",
+        defaultIcon: BiListOl,
+        items: [
+          {
+            type: "item",
+            icon: BiListOl,
+            title: "Ordered list",
+            action: () => editor.chain().focus().toggleOrderedList().run(),
+            isActive: () => editor.isActive("orderedList"),
+          },
+
+          {
+            type: "item",
+            icon: BiListUl,
+            title: "Bullet list",
+            action: () => editor.chain().focus().toggleBulletList().run(),
+            isActive: () => editor.isActive("bulletList"),
+          },
+        ],
+      },
+    ],
+    [editor],
+  )
+
+  return <MenuBar items={items} />
+}
