+++ apps/studio/src/components/PageEditor/MenuBar/index.ts
@@ -1,0 +1,4 @@
+export * from "./TextMenuBar"
+export * from "./AccordionMenuBar"
+export * from "./CalloutMenuBar"
+export * from "./TableMenuBar"
